Started on blending but not happy with it yet.
I like the idea of having a DWIM blend function but I'm not sure about
the other funcs, do I base it on a subset of the GL permutations or do
I based it on blending from art packages? Havent got a good answer so
for now I'll put this down.